feat: correctly defined base inputLevels
